backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
psycopg[binary]==3.2.12
python-dotenv==1.0.1